% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{cellFromLine}
\alias{cellFromLine}
\title{Title}
\usage{
cellFromLine(object, lns)
}
\description{
Title
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
cellFromLine(raster(x), st_cast(x[1:2, ], "MULTILINESTRING"))

}
