% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{cellFromXY}
\alias{cellFromXY}
\title{Title}
\usage{
cellFromXY(object, xy)
}
\value{

}
\description{
Title
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
cellFromXY(raster(x), x)  ## checkout hypertidy/tabularaster
}
