% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{extent.sf}
\alias{extent.sf}
\title{Extent}
\usage{
extent.sf(x)
}
\arguments{
\item{x}{An sf object}
}
\value{
Extent object
}
\description{
This function returns an Extent object of an sf object
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
extent(x)
xmin(x)
xmax(x)
ymin(x)
ymax(x)
}
