% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{rasterize.sf}
\alias{rasterize.sf}
\title{Rasterize an sf object of polygons}
\usage{
rasterize.sf(x, y, ...)
}
\arguments{
\item{x}{an sf object with a geometry of POLYGON or MULTIPOLYGON}

\item{y}{RasterLayer object (raster template)}
}
\value{
A raster of the same size, extent, resolution and projection as the provided raster template
}
\description{
Rasterize set of sf objects (polygon or mulitpolygon)
}
\details{
See \code{\link[fasterize]{fasterize}} for more details.
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
rasterize(x, raster(x))
}
