% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{cellFromPolygon}
\alias{cellFromPolygon}
\title{Title}
\usage{
cellFromPolygon(object, p, weights = FALSE)
}
\description{
Title
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
cellFromPolygon(raster(x, res = 0.1), x[1:4, ])
}
