% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{projection}
\alias{projection}
\title{Title}
\usage{
projection(x, asText = TRUE)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
projection(x)
}
