% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{projection}
\alias{projection}
\title{Get or set a coordinate reference system (projection)}
\usage{
projection(x, asText = TRUE)
}
\arguments{
\item{x}{Raster*, sf, or Spatial object}

\item{...}{}
}
\description{
See [raster::projection()] for more details.
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
projection(x)
}
