% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfraster.R
\name{rasterize.sf}
\alias{rasterize.sf}
\title{Title}
\usage{
rasterize.sf(x, y, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
x <- read_sf(system.file("shape/nc.shp", package="sf"))
rasterize(x, raster(x))
}
